Move TUI to root cmds, organize TUI code
