+++ internal/cmd/policy/policy.go
@@ -16,7 +16,6 @@ import (
 	"github.com/gittuf/gittuf/internal/cmd/policy/removerule"
 	"github.com/gittuf/gittuf/internal/cmd/policy/reorderrules"
 	"github.com/gittuf/gittuf/internal/cmd/policy/sign"
-	"github.com/gittuf/gittuf/internal/cmd/policy/tui"
 	"github.com/gittuf/gittuf/internal/cmd/policy/updateperson"
 	"github.com/gittuf/gittuf/internal/cmd/policy/updaterule"
 	"github.com/gittuf/gittuf/internal/cmd/trustpolicy/apply"
@@ -51,7 +50,6 @@ func New() *cobra.Command {
 	cmd.AddCommand(reorderrules.New(o))
 	cmd.AddCommand(sign.New(o))
 	cmd.AddCommand(stage.New())
-	cmd.AddCommand(tui.New(o))
 	cmd.AddCommand(updateperson.New(o))
 	cmd.AddCommand(updaterule.New(o))
